Put example in one file
